moved and renamed payment method entity test classes #42
